+++ packages/core/rollup.config.js
@@ -1,38 +1,44 @@
-import babel from "rollup-plugin-babel";
-import resolve from "rollup-plugin-node-resolve";
-import commonjs from "rollup-plugin-commonjs";
-import json from "rollup-plugin-json";
-import minify from "rollup-plugin-babel-minify";
-import { name, version, description, author, license } from "./package.json";
+import babel from '@rollup/plugin-babel'
+import resolve from '@rollup/plugin-node-resolve'
+import commonjs from '@rollup/plugin-commonjs'
+import json from '@rollup/plugin-json'
+import { terser } from 'rollup-plugin-terser'
+import { name, version, description, author, license } from './package.json'
 
 export default {
-  input: "src/index.js",
+  input: 'src/index.js',
   output: [
     {
-      file: "dist/index.js",
-      format: "cjs",
-      sourcemap: true
+      file: 'dist/index.js',
+      format: 'cjs',
+      sourcemap: true,
+      exports: 'default'
     },
     {
-      file: "dist/index.mjs",
-      format: "es",
-      sourcemap: true
+      file: 'dist/index.mjs',
+      format: 'es',
+      sourcemap: true,
+      exports: 'default'
     },
     {
-      file: "tests/dist/index.js",
-      format: "cjs",
-      sourcemap: true
+      file: 'tests/dist/index.js',
+      format: 'cjs',
+      sourcemap: true,
+      exports: 'default'
     }
   ],
   plugins: [
+    babel({
+      exclude: 'node_modules/**',
+      babelHelpers: 'bundled'
+    }),
     resolve(),
     json(),
     commonjs(),
-    babel(),
-    minify({
-      comments: false,
-      sourceMap: true,
-      banner: `/**\n * ${name} | v${version}\n * ${description}\n * (c) ${new Date().getFullYear()} ${author}\n * @license ${license}\n */`
+    terser({
+      output: {
+        preamble: `/**\n * ${name} | v${version}\n * ${description}\n * (c) ${new Date().getFullYear()} ${author}\n * @license ${license}\n */`
+      }
     })
   ]
-};
+}