+++ packages/tutorial/rollup.config.js
@@ -1,49 +1,39 @@
-import babel from "rollup-plugin-babel";
-import resolve from "rollup-plugin-node-resolve";
-import commonjs from "rollup-plugin-commonjs";
-import json from "rollup-plugin-json";
-import minify from "rollup-plugin-babel-minify";
-import peerDepsExternal from "rollup-plugin-peer-deps-external";
-import {
-  name,
-  version,
-  description,
-  author,
-  license,
-  main,
-  module
-} from "./package.json";
+//import babel from "@rollup/plugin-babel";
+import resolve from '@rollup/plugin-node-resolve'
+import commonjs from '@rollup/plugin-commonjs'
+import json from '@rollup/plugin-json'
+import { terser } from 'rollup-plugin-terser'
+import peerDepsExternal from 'rollup-plugin-peer-deps-external'
+import { name, version, description, author, license, main, module } from './package.json'
 
 const output = [
   {
     file: main,
-    format: "cjs",
-    sourcemap: true
+    format: 'cjs',
+    sourcemap: true,
+    exports: 'default'
   }
-];
-if (typeof module !== "undefined")
+]
+if (typeof module !== 'undefined')
   output.push({
     file: module,
-    format: "es",
+    format: 'es',
     sourcemap: true
-  });
+  })
 
 export default {
-  input: "src/index.js",
+  input: 'src/index.js',
   output,
   plugins: [
     peerDepsExternal(),
     resolve({ modulesOnly: true }),
     commonjs(),
     json(),
-    babel({
-      exclude: "node_modules/**",
-      plugins: ["@babel/plugin-proposal-object-rest-spread"]
-    }),
-    minify({
-      comments: false,
-      sourceMap: true,
-      banner: `/**\n * ${name} | v${version}\n * ${description}\n * (c) ${new Date().getFullYear()} ${author}\n * @license ${license}\n */`
+    //babel({ exclude: "node_modules/**", }),
+    terser({
+      output: {
+        preamble: `/**\n * ${name} | v${version}\n * ${description}\n * (c) ${new Date().getFullYear()} ${author}\n * @license ${license}\n */`
+      }
     })
   ]
-};
+}